Refactor router widget formatting and fix browser history type
Reformatted code in router widget for improved readability, including multi-line function arguments and match arms. Changed the type of `delta` in `FromWasmBrowserHistoryGo` from `i32` to `f64` to match usage in browser history navigation.

+++ platform/src/os/web/from_wasm.rs
@@ -114,7 +114,7 @@ pub struct FromWasmSetBrowserUrl {
 
 #[derive(FromWasm)]
 pub struct FromWasmBrowserHistoryGo {
-    pub delta: i32,
+    pub delta: f64,
 }
 
 #[derive(FromWasm)]
@@ -348,6 +348,5 @@ pub struct FromWasmStartAudioOutput {
 #[derive(FromWasm)]
 pub struct FromWasmStopAudioOutput {
 }
-
